+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * When mapping a `Condition` without an `asserter`, omit the `Participant` element within the XML.
   Previously this would raise the error "EhrMapperException: Condition.asserter is required" and send a
   failure to the requesting system.
+* When mapping a `MedicationRequest` without a `recorder` or `requester`, omit the `Participant` element within the XML.
+  Previously this would raise the error "MedicationRequest ... missing recorder of type Practitioner, PractitionerRole
+  or Organization" and send a failure to the requesting system.
 
 * When mapping a `DocumentReference` which contains a `NOPAT` `meta.security` or `NOPAT` `securityLabel` tag the resultant XML for that resource
   will contain a `NOPAT` `confidentialityCode` element.